+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 - Updated current index on data series decrease (#121)
 - Updated Image URL to data links (#122)
 - Added slideshow between mixed media (#123)
-- Media panel refactoring (#124)
+- Refactored Media panel (#124)
+- Added multiple queries support in Media Field and Poster Image (#127)
 
 ## 5.1.0 (2024-06-23)